+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Added
 
+- Add a test for updating a repositories which references other authentication repositories. Test repositories are set up programmatically ([386])
+
 ### Changed
 
 - Update find_first_branch_matching_pattern return - only return name of the found branch and not all branches whose name did not match the pattern ([387])
@@ -17,8 +19,11 @@ and this project adheres to [Semantic Versioning][semver].
 
 - Validation of local repositories should not fail if there are no local branches (e.g. after a fresh clone) ([387])
 - Fix GitError exception instantiations ([387])
+- -Fix a minor bug where update status was incorrectly being set in case when a repository with only one commit is cloned ([386])
 
 [387]: https://github.com/openlawlibrary/taf/pull/381
+[386]: https://github.com/openlawlibrary/taf/pull/386
+
 
 ## [0.29.0] - 01/24/2024